requirements.txt
asgiref==3.8.1
Django==5.0.6
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==24.2
django-six==1.0.5
djangorestframework==3.15.1
excel-base==1.0.4
isoweek==1.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.11.0
tzlocal==5.2
xlwt==1.3.0
